+++ src/main/kotlin/com/insyncwithfoo/pyright/configuration/project/ConfigurationPanel.kt
@@ -4,6 +4,7 @@ import com.insyncwithfoo.pyright.configuration.common.ConfigurationPanel
 import com.insyncwithfoo.pyright.message
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.TextFieldWithBrowseButton
+import javax.swing.JCheckBox
 import javax.swing.JLabel
 import javax.swing.JPanel
 
@@ -18,6 +19,8 @@ internal class ConfigurationPanel(private val project: Project) : ConfigurationP
     private lateinit var projectConfigurationFileLabel: JLabel
     private lateinit var projectConfigurationFileInput: TextFieldWithBrowseButton
     
+    private lateinit var autoSuggestExecutableInput: JCheckBox
+    
     override val textFieldsWithBrowseButtons: List<TextFieldWithBrowseButton>
         get() = listOf(projectExecutableInput, projectConfigurationFileInput)
     
@@ -30,11 +33,13 @@ internal class ConfigurationPanel(private val project: Project) : ConfigurationP
     override var configurations: Configurations
         get() = Configurations.create(
             projectExecutable = projectExecutableInput.text.takeIf { it.isNotBlank() },
-            projectConfigurationFile = projectConfigurationFileInput.text.takeIf { it.isNotBlank() }
+            projectConfigurationFile = projectConfigurationFileInput.text.takeIf { it.isNotBlank() },
+            autoSuggestExecutable = autoSuggestExecutableInput.isSelected
         )
         set(value) {
             projectExecutableInput.text = value.projectExecutable.orEmpty()
             projectConfigurationFileInput.text = value.projectConfigurationFile.orEmpty()
+            autoSuggestExecutableInput.isSelected = value.autoSuggestExecutable
         }
     
     override fun getService() = ConfigurationService.getInstance(project)
@@ -42,6 +47,7 @@ internal class ConfigurationPanel(private val project: Project) : ConfigurationP
     override fun setLabels() {
         projectExecutableLabel.text = message("configurations.project.projectExecutable.label")
         projectConfigurationFileLabel.text = message("configurations.project.projectConfigurationFile.label")
+        autoSuggestExecutableInput.text = message("configurations.project.autoSuggestExecutable.label")
     }
     
 }